# Perf improvements part 2