--- a/src/System.CommandLine/Option.cs
+++ b/src/System.CommandLine/Option.cs
@@ -18,7 +18,6 @@ namespace System.CommandLine
         IOption
     {
         private string? _name;
-        private protected readonly HashSet<string> _unprefixedAliases = new();
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Option"/> class.
@@ -206,10 +205,6 @@ namespace System.CommandLine
             ThrowIfAliasIsInvalid(alias);
 
             base.AddAliasInner(alias);
-
-            var unprefixedAlias = alias.RemovePrefix();
-
-            _unprefixedAliases.Add(unprefixedAlias!);
         }
 
         /// <summary>
@@ -223,13 +218,19 @@ namespace System.CommandLine
         /// </summary>
         /// <param name="alias">The alias, which can include a prefix.</param>
         /// <returns><see langword="true"/> if the alias exists; otherwise, <see langword="false"/>.</returns>
-        public bool HasAliasIgnoringPrefix(string alias) => _unprefixedAliases.Contains(alias.RemovePrefix());
-
-        private protected override void RemoveAlias(string alias)
+        public bool HasAliasIgnoringPrefix(string alias)
         {
-            _unprefixedAliases.Remove(alias);
+            ReadOnlySpan<char> rawAlias = alias.AsSpan(alias.GetPrefixLength());
 
-            base.RemoveAlias(alias);
+            foreach (string existingAlias in _aliases)
+            {
+                if (MemoryExtensions.Equals(existingAlias.AsSpan(existingAlias.GetPrefixLength()), rawAlias, StringComparison.CurrentCulture))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         /// <summary>