|        Method |      Args | Mean before | Mean after | Allocated before | Allocated after |
|-------------- |---------- |------------:|-----------:|-----------------:|----------------:|
|  DefaultsSync | String[0] |   45.093 us |  38.805 us |            32 KB |           25 KB |
| DefaultsAsync | String[0] |   45.541 us |  38.980 us |            32 KB |           25 KB |
|   MinimalSync | String[0] |    8.666 us |   6.903 us |            14 KB |           12 KB |
|  MinimalAsync | String[0] |    8.946 us |   7.146 us |            14 KB |           12 KB |
|  DefaultsSync | String[4] |   50.876 us |  43.782 us |            37 KB |           29 KB |
| DefaultsAsync | String[4] |   51.021 us |  44.395 us |            37 KB |           29 KB |
|   MinimalSync | String[4] |   12.436 us |  10.294 us |            18 KB |           16 KB |
|  MinimalAsync | String[4] |   12.603 us |  10.698 us |            18 KB |           16 KB |

cc @jonsequitur 

I'll most probably send a few more PRs, I am currently focusing on things that don't introduce breaking changes